PDCT-716 revert errors move
