+++ app/api/api_v1/routers/analytics.py
@@ -1,7 +1,7 @@
 """Endpoints for managing the Analytics service."""
 import logging
 from fastapi import APIRouter, HTTPException, status
-from app.clients.db.errors import RepositoryError
+from app.errors import RepositoryError
 
 import app.service.analytics as analytics_service
 from app.model.analytics import SummaryDTO
