+++ app/repository/document.py
@@ -24,7 +24,7 @@ from db_client.models.law_policy.family import (
     DocumentStatus,
     Slug,
 )
-from app.clients.db.errors import RepositoryError, ValidationError
+from app.errors import RepositoryError, ValidationError
 from app.model.document import DocumentCreateDTO, DocumentReadDTO, DocumentWriteDTO
 from app.repository import family as family_repo
 from app.repository.helpers import generate_import_id, generate_slug
