+++ app/service/app_user.py
@@ -1,5 +1,5 @@
 from sqlalchemy.orm import Session
-from app.clients.db.errors import ValidationError
+from app.errors import ValidationError
 from app.repository import app_user_repo
