+++ app/service/geography.py
@@ -1,5 +1,5 @@
 from sqlalchemy.orm import Session
-from app.clients.db.errors import ValidationError
+from app.errors import ValidationError
 from app.repository import geography_repo
