+++ integration_tests/mocks/bad_document_repo.py
@@ -1,6 +1,6 @@
 from typing import Optional
 from pytest import MonkeyPatch
-from app.clients.db.errors import RepositoryError
+from app.errors import RepositoryError
 
 from app.model.document import DocumentCreateDTO, DocumentReadDTO
