+++ unit_tests/mocks/repos/collection_repo.py
@@ -3,7 +3,7 @@ from typing import Optional
 from pytest import MonkeyPatch
 from sqlalchemy import exc
 
-from app.clients.db.errors import RepositoryError
+from app.errors import RepositoryError
 from app.model.collection import CollectionReadDTO
 from unit_tests.helpers.collection import create_collection_read_dto
