+++ unit_tests/mocks/services/analytics_service.py
@@ -1,5 +1,5 @@
 from pytest import MonkeyPatch
-from app.clients.db.errors import RepositoryError
+from app.errors import RepositoryError
 
 from app.model.analytics import SummaryDTO
 from unit_tests.helpers.analytics import create_summary_dto
