+++ unit_tests/mocks/services/config_service.py
@@ -1,5 +1,5 @@
 from pytest import MonkeyPatch
-from app.clients.db.errors import RepositoryError
+from app.errors import RepositoryError
 
 from app.model.config import ConfigReadDTO, DocumentConfig, EventConfig
