+++ unit_tests/mocks/services/family_service.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from pytest import MonkeyPatch
 
-from app.clients.db.errors import RepositoryError
+from app.errors import RepositoryError
 from app.model.family import FamilyReadDTO, FamilyWriteDTO
 from unit_tests.helpers.family import create_family_dto
