+++ unit_tests/service/test_analytics_service.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 import pytest
 
-from app.clients.db.errors import RepositoryError
+from app.errors import RepositoryError
 from app.model.analytics import SummaryDTO
 import app.service.analytics as analytics_service
