+++ unit_tests/service/test_authorisation_service.py
@@ -1,5 +1,5 @@
 import pytest
-from app.clients.db.errors import AuthenticationError, RepositoryError
+from app.errors import AuthenticationError, RepositoryError
 import app.service.authentication as auth_service
 import app.service.token as token_service
 from unit_tests.mocks.repos.app_user_repo import (
