+++ unit_tests/service/test_event_service.py
@@ -1,7 +1,7 @@
 import pytest
 
 import app.service.event as event_service
-from app.clients.db.errors import RepositoryError, ValidationError
+from app.errors import RepositoryError, ValidationError
 from app.model.event import EventReadDTO, EventWriteDTO
 from unit_tests.helpers.event import create_event_create_dto
