+++ unit_tests/service/test_id_service.py
@@ -1,5 +1,5 @@
 import pytest
-from app.clients.db.errors import ValidationError
+from app.errors import ValidationError
 from app.service import id
